euraxess/app/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
httpx==0.24.0
python-dateutil==2.8.2
xmltodict==0.13.0
beautifulsoup4==4.12.2
